+++ BrightID/src/components/Groups/GroupCard.tsx
@@ -6,7 +6,7 @@ import { DEVICE_TYPE } from '@/utils/deviceConstants';
 import { DARK_ORANGE, WHITE, DARK_GREEN, DARK_GREY } from '@/theme/colors';
 import { fontSize } from '@/theme/fonts';
 import { getGroupName, ids2connections } from '@/utils/groups';
-import GroupPhoto from './GroupPhoto';
+import { GroupPhoto } from '@/components/Groups/GroupPhoto';
 
 /**
  * Group Card in the Groups Screen
