+++ BrightID/src/routes/Groups.tsx
@@ -10,8 +10,8 @@ import { fontSize } from '@/theme/fonts';
 import { GroupsScreen } from '@/components/Groups/GroupsScreen';
 import SearchGroups from '@/components/Helpers/SearchGroups';
 import SearchConnections from '@/components/Helpers/SearchConnections';
-import NewGroupScreen from '@/components/Groups/NewGroups/NewGroupScreen';
-import GroupInfoScreen from '@/components/Groups/NewGroups/GroupInfoScreen';
+import { NewGroupScreen } from '@/components/Groups/NewGroups/NewGroupScreen';
+import { GroupInfoScreen } from '@/components/Groups/NewGroups/GroupInfoScreen';
 import { MembersScreen } from '@/components/Groups/Members/MembersScreen';
 import ConnectionsScreen from '@/components/Connections/ConnectionsScreen';
 import { headerOptions, NavHome, AnimatedHeaderTitle } from './helpers';
